fix(compare): share field-value frequency rollup with compare subcommand
The cross-file field-value frequency rollup added in #1106 lived only in the
xtask `compare-outputs` tool, so the `provenant compare` subcommand emitted
per-file diffs without the cross-file frequency view.

Move the rollup (`field_value_frequency_rollup`, the
`FieldValueFrequencyEntry`/`FieldValueFrequencyDirections` types,
`field_value_frequency_summary`, and the two top-N consts) into the shared
`compare_driver_shared` module both drivers already depend on, and wire it into
`compare::write_comparison_artifacts` so the subcommand now writes
`samples/field_value_frequency.json` and a `field_value_frequency_top` preview
in `summary.json`. The xtask tool consumes the shared definitions with
byte-identical output.

The rollup stays a pure diagnostic: it never feeds `comparison_status` or any
signal count, and keeps the neutral PV-only / SC-only framing from #1106.
